SmartSDLC_HF/requirements.txt
streamlit
python-dotenv
requests
pymupdf